+++ examples/hello_world_2.py
@@ -11,16 +11,16 @@ d = DaisyUI(g)
 # callbacks
 async def hello_button_click(button: Component, event: Event):
     v: int = randint(0, 100)
-    hello_card.content = f'Hello {v}'
-    hello_button.content = f'Hello {v}'
+    hello_card.data = f'Hello {v}'
+    hello_button.data = f'Hello {v}'
 
 async def world_button_click(button: Component, event: Event):
     v: int = randint(0, 100)
-    world_card.content = f'World {v}'
-    world_button.content = f'World {v}'
+    world_card.data = f'World {v}'
+    world_button.data = f'World {v}'
 
 page = h.Page()
-page.head.add(title := h.Title(content='Hello world 1'))
+page.head.add(title := h.Title(data='Hello world 1'))
 page.head.add(link := h.Link(rel='shortcut icon', type='image/png', href='/static/gladius/favicon.png'))
 page.head.add(link := h.Link(href='https://cdn.jsdelivr.net/npm/daisyui@3.8.0/dist/full.css', rel='stylesheet', type='text/css'))
 page.head.add(script := h.Script(src='https://cdn.tailwindcss.com'))